+++ pde/src/sgpp/pde/application/HeatEquationSolver.cpp
@@ -34,7 +34,7 @@ HeatEquationSolver::~HeatEquationSolver() {
   }
 }
 
-void HeatEquationSolver::constructGrid(base::BoundingBox& BoundingBox, int level) {
+void HeatEquationSolver::constructGrid(base::BoundingBox& BoundingBox, size_t level) {
   this->dim = BoundingBox.getDimensions();
   this->levels = level;
